Feat: Calc Room Total Cost
solved #95
Calc cost to verify payment

+++ src/main/java/com/buildup/kbnb/controller/PaymentController.java
@@ -37,7 +37,7 @@ public class PaymentController {
     @NoArgsConstructor
     public static class PaymentDto {
         private String receipt_id;
-        private Integer price;
+        private Double price;
     }
 
     @PostMapping("/cancel")
